+++ R/kfold.R
@@ -35,6 +35,9 @@
 #'   the returned object to store the cross-validated \code{brmsfit} 
 #'   objects and the indices of the omitted observations for each fold. 
 #'   Defaults to \code{FALSE}.
+#' @param future_args A list of further arguments passed to
+#'   \code{\link[future:future]{future}} for additional control over parallel
+#'   execution if activated.
 #'   
 #' @return \code{kfold} returns an object that has a similar structure as the 
 #'   objects returned by the \code{loo} and \code{waic} methods and
@@ -98,7 +101,8 @@
 #' @export
 kfold.brmsfit <- function(x, ..., K = 10, Ksub = NULL, folds = NULL, 
                           group = NULL, exact_loo = NULL, compare = TRUE,
-                          resp = NULL, model_names = NULL, save_fits = FALSE) {
+                          resp = NULL, model_names = NULL, save_fits = FALSE,
+                          future_args = list()) {
   args <- split_dots(x, ..., model_names = model_names)
   use_stored <- ulapply(args$models, function(x) is_equal(x$kfold$K, K))
   if (!is.null(exact_loo) && as_one_logical(exact_loo)) {
@@ -107,7 +111,7 @@ kfold.brmsfit <- function(x, ..., K = 10, Ksub = NULL, folds = NULL,
   }
   c(args) <- nlist(
     criterion = "kfold", K, Ksub, folds, group, 
-    compare, resp, save_fits, use_stored
+    compare, resp, save_fits, future_args, use_stored
   )
   do_call(compute_loolist, args)
 }
@@ -116,9 +120,9 @@ kfold.brmsfit <- function(x, ..., K = 10, Ksub = NULL, folds = NULL,
 # @inheritParams kfold.brmsfit
 # @param model_name ignored but included to avoid being passed to '...'
 .kfold <- function(x, K, Ksub, folds, group, save_fits,
-                   newdata, resp, model_name, 
+                   newdata, resp, model_name, future_args = list(),
                    newdata2 = NULL, ...) {
-  stopifnot(is.brmsfit(x))
+  stopifnot(is.brmsfit(x), is.list(future_args))
   if (is.brmsfit_multiple(x)) {
     warn_brmsfit_multiple(x)
     class(x) <- "brmsfit"
@@ -239,12 +243,13 @@ kfold.brmsfit <- function(x, ..., K = 10, Ksub = NULL, folds = NULL,
   }
   
   x <- recompile_model(x)
+  future_args$FUN <- .kfold_k
+  future_args$seed <- TRUE
   for (k in Ksub) {
     ks <- match(k, Ksub)
     message("Fitting model ", k, " out of ", K)
-    futures[[ks]] <- future::future(
-      .kfold_k(k), seed = TRUE
-    )
+    future_args$args <- list(k)
+    futures[[ks]] <- do_call("futureCall", future_args, pkg = "future")
   }
   for (k in Ksub) {
     ks <- match(k, Ksub)